bug fixes around cloning threads, properly set the organization ID and the discriminator. flag case owners in the project assignment DTO, used to locking case owner permissions. introduced a new role for authorized staff, staff members who have access to connection requests. refactored the GET method in the FormTemplateService and separated it into three functions for each use case. OrgAdmins no longer need an assignment to access projects, form templates and records, they have full access to everything under their own organization. when users move between organizations, remove their current assignments, and unassign clients/staff from their cases. it's no longer necessary to assign root users to users since OrgAdmins don't need assignments anymore. refactored the SurveyService, OrgAdmins don't need assignments, and when OrgUsers request aggregate cases exclude records posted by other users. only staff members can join teams, and in the management wizard, only clients can be assigned to staff members (we don't allow client-to-client connections). installed angular-wizard and particles.js locally and removed the CDN references. updated hangfire packages. updated HangFireConfig with CRON expressions. users can now remove their security question and answer. when new clients sign up, it's no longer necessary to assign the OnRecord admin. fixed an authorization bug in the Feedbacks controller, the SendEmail method is available to all admins. when users are removed from an organization, invalidate cache entries. users who have previously sent connection requests, can request to join organizations again. when deleting connection reqeusts, only notify OrgUsers if the request hasn't been approved. when connection requests are approved, invalidate projects from the cache. fixed a bug in the OrgTeams controller, when notifying users about leaving. added a new method to the OrgUsers controller for getting staff members for an organization. OrgAdmins are not allowed to create clients, requests will be rejected by a 401 response. when new OrgAdmins are created, it's no longer necessary to assign all organization projects to them, OrgAdmins don't need assignments. when updating users, system admins can change emails/username and the current project, but OrgAdmins can only change emails/username for staff members. added a new method to the OrgUsers controller for adding/removing staff members to and from the authorized staff role. when OrgAdmins create new projects (aggregate cases) there's no need to create an assignment for them. client-to-client assignments are rejected by the API. when exporting pdf/zip files, OrgAdmins don't need assignments but they only have access to their organization. developed a new method for downloading account data, a multi-part zip file will be generated along with JSON files containing textual data. an email is sent with download links, and the generated data will expire automatically and be wiped from the server (experimental). refactored the SurveysController, OrgAdmins no longer need assignments, and the GetUserSurveys method has been rewritten to correctly collect records and advice records, and to exclude records posted by other users when requesting aggregate cases. same thing for the Search method. refactored the code for parsing serial references, now using RegEx patterns to accurately extract serial numbers. laid the foundation for switching color themes, currently disabled. updated all smart-tables and added pagination controls at the top as well. in surveys and advices views, select the current project by default. in the calendar view, if there are no records to show, display an info alert. updated projects dropdowns to include emails in the template. users can now unlink from their organization in the subscriptions view. turned the security questions dropdown into an auto-complete. display case names in the create-thread and create-advice-thread modals. added counters to selection buttons in the project assignments view, refactored the TypeScript controller to reload data when entries change, and case owner permissions are locked so that we don't accidentally change them. added a new modal to the summary page for listing advice references. charts and locations are excluded from advice summaries. fixed a bug in the export view, clicking the reset button should bring back removed attachments. disabled shared-threads (not used for now). added counters to selection buttons in the statistics view. surveys routes now resolve the project using state params, to set titles and breadcrumbs accordingly. removed the metric-group header from the template. disallow duplicate items in the serial-references directive. numerous general bug fixes and improvements. refactoring and code cleanup will follow in the next commit.

+++ Models/Services/SubscriptionService.cs
@@ -262,14 +262,26 @@ namespace LightMethods.Survey.Models.Services
 
         public void MoveUserToOrganization(Organisation organisation, OrgUser orgUser)
         {
-            // remove this user from any teams in current organisation.
+            // remove this user from any teams in the current organisation.
             RemoveUserFromCurrentTeams(orgUser);
+
+            // remove this user from any assignments in the current organization.
+            RemoveUserFromCurrentAssignments(orgUser);
+
+            // remove the assignment from current organisation's root user
+            // NOTE: this could be removed too. we're not assigning root users
+            // so there's no need to unassign them. doesn't hurt to confirm though.
+            UnassignRootUserFromProject(orgUser, orgUser.Organisation);
+
+            // remove any staff/client assignments from current organization
+            UnassignStaffAndClientsFromProject(orgUser, orgUser.Organisation);
 
             // move the user and their personal case to the new organisation.
             MoveUserCaseToOrganization(orgUser, organisation);
 
             // assign Org root user to the project
-            AssignRootUserToProject(orgUser.CurrentProjectId.Value, organisation.RootUserId.Value);
+            // NOTE: not necessary, because OrgAdmins don't need assignments.
+            //AssignRootUserToProject(orgUser.CurrentProjectId.Value, organisation.RootUserId.Value);
 
             // close last subscription
             CloseLastSubscription(orgUser);
@@ -293,24 +305,49 @@ namespace LightMethods.Survey.Models.Services
             // remove user from any teams in current organisation.
             RemoveUserFromCurrentTeams(orgUser);
 
-            // move user back to OnRecord
-            MoveUserCaseToOrganization(orgUser, OnRecord);
+            // remove this user from any assignments in the current organization.
+            RemoveUserFromCurrentAssignments(orgUser);
 
             // remove the assignment from current organisation's root user
-            UnassignRootUserFromProject(orgUser, organisation, OnRecord);
+            UnassignRootUserFromProject(orgUser, organisation);
+
+            // remove any staff/client assignments from current organization
+            UnassignStaffAndClientsFromProject(orgUser, organisation);
+
+            // assign OnRecord admin to user's project
+            // NOTE: not necessary, since OrgAdmins don't need assignments.
+            //if (OnRecord.RootUser != null)
+            //    AssignRootUserToProject(orgUser.CurrentProjectId.Value, OnRecord.RootUser.Id);
+
+            // move user back to OnRecord
+            MoveUserCaseToOrganization(orgUser, OnRecord);
 
             // close the last organisation subscription.
             CloseLastOrganisationSubscription(orgUser);
 
             // subscribe user to OnRecord again.
             SubscribeUserToOrganization(orgUser, OnRecord);
+
+            // NOTE: I think we should disable export permissions here.
+            // and also set IsSubscribed = false.
         }
 
         #endregion
 
         #region Helpers
 
-        public void UnassignRootUserFromProject(OrgUser user, Organisation organisation, Organisation OnRecord)
+        public void RemoveUserFromCurrentAssignments(OrgUser user)
+        {
+            var assignments = UnitOfWork.AssignmentsRepository
+                .AllAsNoTracking
+                .Where(x => x.OrgUserId == user.Id && x.ProjectId != user.CurrentProjectId)
+                .Where(x => x.Project.OrganisationId == user.OrganisationId)
+                .ToList();
+
+            UnitOfWork.AssignmentsRepository.Delete(assignments);
+        }
+
+        public void UnassignRootUserFromProject(OrgUser user, Organisation organisation)
         {
             if (user.CurrentProject != null && user.CurrentProject.CreatedById == user.Id)
             {
@@ -318,39 +355,52 @@ namespace LightMethods.Survey.Models.Services
                 var rootUserAssignment = UnitOfWork.AssignmentsRepository
                     .AllAsNoTracking
                     .Where(x => x.OrgUserId == organisation.RootUserId && x.ProjectId == user.CurrentProject.Id)
-                    .FirstOrDefault();
+                    .SingleOrDefault();
 
-                UnitOfWork.AssignmentsRepository.Delete(rootUserAssignment);
+                if (rootUserAssignment != null)
+                    UnitOfWork.AssignmentsRepository.Delete(rootUserAssignment);
+            }
+        }
 
-                // assign project to OnRecord admin again
-                if (OnRecord.RootUser != null)
-                {
-                    var OnRecordRootAssignment = UnitOfWork.AssignmentsRepository
-                        .AllAsNoTracking
-                        .Where(x => x.OrgUserId == OnRecord.RootUserId && x.ProjectId == user.CurrentProject.Id)
-                        .FirstOrDefault();
+        public void UnassignStaffAndClientsFromProject(OrgUser user, Organisation organisation)
+        {
+            if (user.CurrentProject != null && user.CurrentProject.CreatedById == user.Id)
+            {
+                var assignments = UnitOfWork.AssignmentsRepository
+                    .AllAsNoTracking
+                    .Where(x => x.ProjectId == user.CurrentProject.Id)
+                    .Where(x => x.OrgUserId != user.Id)
+                    .Where(x => !x.OrgUser.IsRootUser)
+                    .Where(x => x.OrgUser.Organisation.Id == organisation.Id)
+                    .ToList();
 
-                    if (OnRecordRootAssignment == null)
-                        AssignRootUserToProject(user.CurrentProject.Id, OnRecord.RootUser.Id);
-                }
+                UnitOfWork.AssignmentsRepository.Delete(assignments);
             }
         }
 
         public void AssignRootUserToProject(Guid projectId, Guid rootUserId)
         {
-            var rootUserAssignment = new Assignment
-            {
-                ProjectId = projectId,
-                OrgUserId = rootUserId,
-                CanView = true,
-                CanAdd = true,
-                CanEdit = true,
-                CanDelete = true,
-                CanExportPdf = true,
-                CanExportZip = true
-            };
+            var assignment = UnitOfWork.AssignmentsRepository
+                .AllAsNoTracking
+                .Where(x => x.ProjectId == projectId && x.OrgUserId == rootUserId)
+                .SingleOrDefault();
 
-            UnitOfWork.AssignmentsRepository.InsertOrUpdate(rootUserAssignment);
+            if (assignment == null)
+            {
+                var rootUserAssignment = new Assignment
+                {
+                    ProjectId = projectId,
+                    OrgUserId = rootUserId,
+                    CanView = true,
+                    CanAdd = true,
+                    CanEdit = true,
+                    CanDelete = true,
+                    CanExportPdf = true,
+                    CanExportZip = true
+                };
+
+                UnitOfWork.AssignmentsRepository.InsertOrUpdate(rootUserAssignment);
+            }
         }
 
         public void RemoveUserFromCurrentTeams(OrgUser user)
